backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pydantic
requests
razorpay
python-dotenv
APScheduler==3.10.4
